pyproject.toml
[project]
name = "redir-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "configparser>=7.2.0",
    "discord-py>=2.7.1",
    "fastapi[standard]>=0.136.1",
    "requests>=2.33.1",
]
